requirements.txt
aiogram
fastapi
uvicorn
python-dotenv
asyncpg
stripe